+++ Countform/countform.php
@@ -1,0 +1,8 @@
+<?php
+require "functions.php";
+
+showHTML();
+
+if (!is_null($_POST['usernumber'] ?? null)) {
+  guessNumber();
+}
